+++ ui/main/game/galactic_war/cards/gwc_start_orbital.js
@@ -35,7 +35,7 @@ define([
               file: gwoUnit.orbitalFabber,
               path: "buildable_types",
               op: "add",
-              value: " | FabBuild",
+              value: " | FabBuild & Custom58",
             },
           ]);
         } else {